set 1, challenge 4
Let set_1_challenge_3 take a string arugment instead of hard-coding the
ciphertext. Then, write a shell script for set_1_challenge_4 that uses
the set_1_challenge_3 program to find the most likely English strings in
a file. I found the string through a combination of tweaking the scoring
function and brute force :)